created email facade
