+++ backend/src/model/Model.ts
@@ -1,7 +1,7 @@
 import User from "./User.js";
 import DatabaseFacade from "./facades/DatabaseFacade.js";
 import SpotifyFacade from "./facades/SpotifyFacade.js";
-import { MailerSend, EmailParams, Recipient, Sender } from "mailersend";
+import EmailFacade from "./facades/EmailFacade.js";
 
 /**
  * The Model class is responsible for modifying and fetching backend data
@@ -9,9 +9,7 @@ import { MailerSend, EmailParams, Recipient, Sender } from "mailersend";
 class Model {
     private db: DatabaseFacade;
     private spotify: SpotifyFacade;
-    private readonly mailer = new MailerSend({
-        apiKey: process.env.MAILER_KEY || "",
-    });
+    private email: EmailFacade;
 
     /**
      * @effects constructs facades
@@ -19,6 +17,7 @@ class Model {
     constructor() {
         this.db = new DatabaseFacade();
         this.spotify = new SpotifyFacade();
+        this.email = new EmailFacade();
     }
 
     /**
@@ -61,7 +60,7 @@ class Model {
         user = await this.generatePlaylist(user, true);
         if (notifs == "true") {
             user.notifs = true;
-            this.sendWelcomeEmail(user);
+            this.email.sendWelcomeEmail(user);
             this.db.updateUser(user);
         } else {
             user.notifs = false;
@@ -90,53 +89,8 @@ class Model {
         let users: [User] = await this.db.getOptedInUsers();
         for (let i = 0; i < users.length; i++) {
             users[i] = await this.generatePlaylist(users[i], false);
-            this.sendNewPlaylistEmail(users[i]);
+            this.email.sendNewPlaylistEmail(users[i]);
         }
-    }
-
-    /**
-     * @param user the user to email
-     * @effects sends a welcome email to the provided user's email.
-     */
-    async sendWelcomeEmail(user: User): Promise<void> {
-        if (!user.email) throw new Error("User must have email");
-        const sentFrom = new Sender(process.env.EMAIL || "", "Vaultify");
-        const recipients = [new Recipient(user.email)];
-        const emailParams = new EmailParams()
-            .setFrom(sentFrom)
-            .setTo(recipients)
-            .setSubject("Welcome to Vaultify!")
-            .setText(
-                "Congratulations on creating your first playlist with Vaultify!\nWe'll be keeping touch every month with a brand new playlist for you to enjoy!\n\nSincerely,\nThe Vaultify Team :)"
-            );
-
-        this.mailer.email
-            .send(emailParams)
-            .then((response) => console.log(response))
-            .catch((error) => console.log(error));
-    }
-
-    /**
-     * @param user the user to email
-     * @effects sends an email with a link to the user's new playlist
-     */
-    async sendNewPlaylistEmail(user: User): Promise<void> {
-        if (!user.email || !user.playlists || user.playlists.length == 0) {
-            throw new Error("User must have email and playlists");
-        }
-        const sentFrom = new Sender(process.env.EMAIL || "", "Vaultify");
-        const recipients = [new Recipient(user.email)];
-        const emailParams = new EmailParams()
-            .setFrom(sentFrom)
-            .setTo(recipients)
-            .setSubject("Your Monthly Recap is Here!")
-            .setText(
-                `Here's the link to your monthly recap:\n${user.playlists[user.playlists.length - 1]}\n\nEnjoy!\n-The Vaultify Team :)`
-            );
-
-        this.mailer.email
-            .send(emailParams)
-            .catch((error) => console.log(error));
     }
 }
